fix(speed): improve speed precision, deduplication logic, and idle search

+++ idle-monitor/app/Models/GpsTrack.php
@@ -41,7 +41,7 @@ class GpsTrack extends Model
     protected $casts = [
         'longitude' => 'decimal:7',
         'latitude' => 'decimal:7',
-        'speed' => 'integer',
+        'speed' => 'float',
         'gps_time' => 'datetime',
         'report_time' => 'datetime',
         'is_acc_on' => 'boolean',
